Explicitly set real literals to float before adding to message box.

+++ main.cpp
@@ -39,11 +39,11 @@ int main()
         intMessageBox.send(3, 4);
         intMessageBox.send(5, 5);
 
-        floatMessageBox.send(0, 1.1);
-        floatMessageBox.send(1, 2.2);
-        floatMessageBox.send(2, 0.0);
-        floatMessageBox.send(3, 4.4);
-        floatMessageBox.send(6, 5.5);
+        floatMessageBox.send(0, (float) 1.1);
+        floatMessageBox.send(1, (float) 2.2);
+        floatMessageBox.send(2, (float) 0.0);
+        floatMessageBox.send(3, (float) 4.4);
+        floatMessageBox.send(6, (float) 5.5);
 
         std::cout << std::endl
                   << "Testing toString() method and operator overloading."
